// mobile-app/types/Apartment.ts
type Apartment = {
  _id: string;
  thumbnail: string;
  price: number;
  residenceType: string;
  compound: string;
  noOfRooms: number;
  noOfBaths: number;
  size: number;
  imageGallery: string[];
  startPrice: number;
  developer: string;
  finishing: string;
  paymentPlan: string;
  paymentDuration: number;
};

export default Apartment;